+++ src/components/jobCardOrderPage.js
@@ -1,4 +1,3 @@
-import boat from "../images/boat.jpg";
 import "../app.css";
 
 export default function JobCardOrderPage({
@@ -21,11 +20,18 @@ export default function JobCardOrderPage({
       style={{ height: isSelected ? "auto" : "fit-content" }}
     >
       <div className="col-span-1">
-        <img
-          className="h-full w-full object-cover"
-          src={img ? img : boat}
-          alt={job.fileName}
-        />
+      <div className="h-full w-full object-cover">
+        {(img)? 
+          (<img            
+            src={img}
+            alt={job.fileName}
+          /> )
+          :
+          (<svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" strokeWidth={1.5} stroke="gray">
+            <path strokeLinecap="round" strokeLinejoin="round" d="M18.364 18.364A9 9 0 0 0 5.636 5.636m12.728 12.728A9 9 0 0 1 5.636 5.636m12.728 12.728L5.636 5.636" />
+          </svg>)
+        } 
+      </div>
       </div>
       <div className="col-span-2 pl-5 pt-5 text-gray-600 w-full">
         <div className="uppercase tracking-wide text-sm text-indigo-500 font-semibold">
